Refactor classes to packages
