+++ src/main/java/noob/exception/NoobException.java
@@ -1,4 +1,4 @@
-package noob;
+package noob.exception;
 
 public class NoobException extends Exception {
     public NoobException(String msg) {
