+++ src/main/java/noob/task/TaskType.java
@@ -1,4 +1,4 @@
-package noob;
+package noob.task;
 
 public enum TaskType {
     DEADLINE, TODO, EVENT
